+++ sdk/computelimit/azure-resourcemanager-computelimit/src/main/java/com/azure/resourcemanager/computelimit/ComputeLimitManager.java
@@ -26,9 +26,11 @@ import com.azure.core.util.CoreUtils;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.resourcemanager.computelimit.fluent.ComputeLimitClient;
 import com.azure.resourcemanager.computelimit.implementation.ComputeLimitClientBuilder;
+import com.azure.resourcemanager.computelimit.implementation.FeaturesImpl;
 import com.azure.resourcemanager.computelimit.implementation.GuestSubscriptionsImpl;
 import com.azure.resourcemanager.computelimit.implementation.OperationsImpl;
 import com.azure.resourcemanager.computelimit.implementation.SharedLimitsImpl;
+import com.azure.resourcemanager.computelimit.models.Features;
 import com.azure.resourcemanager.computelimit.models.GuestSubscriptions;
 import com.azure.resourcemanager.computelimit.models.Operations;
 import com.azure.resourcemanager.computelimit.models.SharedLimits;
@@ -50,6 +52,8 @@ public final class ComputeLimitManager {
     private GuestSubscriptions guestSubscriptions;
 
     private SharedLimits sharedLimits;
+
+    private Features features;
 
     private final ComputeLimitClient clientObject;
 
@@ -300,6 +304,18 @@ public final class ComputeLimitManager {
             this.sharedLimits = new SharedLimitsImpl(clientObject.getSharedLimits(), this);
         }
         return sharedLimits;
+    }
+
+    /**
+     * Gets the resource collection API of Features.
+     * 
+     * @return Resource collection API of Features.
+     */
+    public Features features() {
+        if (this.features == null) {
+            this.features = new FeaturesImpl(clientObject.getFeatures(), this);
+        }
+        return features;
     }
 
     /**