+++ sdk/computelimit/azure-resourcemanager-computelimit/src/main/java/com/azure/resourcemanager/computelimit/fluent/ComputeLimitClient.java
@@ -66,4 +66,11 @@ public interface ComputeLimitClient {
      * @return the SharedLimitsClient object.
      */
     SharedLimitsClient getSharedLimits();
+
+    /**
+     * Gets the FeaturesClient object to access its operations.
+     * 
+     * @return the FeaturesClient object.
+     */
+    FeaturesClient getFeatures();
 }