+++ sdk/computelimit/azure-resourcemanager-computelimit/CHANGELOG.md
@@ -1,14 +1,24 @@
 # Release History
 
-## 1.0.0-beta.2 (Unreleased)
+## 1.0.0 (2026-03-31)
+
+- Azure Resource Manager ComputeLimit client library for Java. This package contains Microsoft Azure SDK for ComputeLimit Management SDK. Microsoft Azure Compute Limit Resource Provider. Package api-version 2026-03-20. For documentation on how to use this package, please see [Azure Management Libraries for Java](https://aka.ms/azsdk/java/mgmt).
 
 ### Features Added
 
-### Breaking Changes
+* `models.OperationStatusResult` was added
 
-### Bugs Fixed
+* `models.FeatureProperties` was added
 
-### Other Changes
+* `models.Feature` was added
+
+* `models.FeatureState` was added
+
+* `models.Features` was added
+
+#### `ComputeLimitManager` was modified
+
+* `features()` was added
 
 ## 1.0.0-beta.1 (2025-11-12)